:bug: QD-14804 Make chmod-based fs tests fail loudly when DAC is bypassed (root/privileged container)
