+++ internal/foundation/fs/canonical_execonly_test.go
@@ -7,14 +7,25 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/JetBrains/qodana-cli/internal/testutil/needs"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func TestCanonical_ExecOnlyDir(t *testing.T) {
-	if os.Getuid() == 0 {
-		t.Skip("root bypasses permission checks")
+// requireExecOnlyEnforced verifies that chmodding `dir` to 0o111 (execute-only)
+// actually denies ReadDir. Root or CAP_DAC_READ_SEARCH bypass DAC and make these
+// tests meaningless — surface that loudly with an actionable opt-out hint.
+func requireExecOnlyEnforced(t *testing.T, dir string) {
+	t.Helper()
+	if _, err := os.ReadDir(dir); err == nil {
+		t.Fatalf("environment does not enforce 0o111 directory perms "+
+			"(uid=%d, gid=%d); run as non-root user, or set %s=0 to skip",
+			os.Getuid(), os.Getgid(), needs.NonRoot.EnvVar)
 	}
+}
+
+func TestCanonical_ExecOnlyDir(t *testing.T) {
+	needs.Need(t, needs.NonRoot)
 
 	tmp := canonicalTempDir(t)
 	dir := filepath.Join(tmp, "execonly")
@@ -23,6 +34,7 @@ func TestCanonical_ExecOnlyDir(t *testing.T) {
 
 	require.NoError(t, os.Chmod(dir, 0o111))
 	t.Cleanup(func() { _ = os.Chmod(dir, 0o755) })
+	requireExecOnlyEnforced(t, dir)
 
 	got, err := Canonical(filepath.Join(dir, "file"))
 	require.NoError(t, err, "Canonical must work with execute-only parent dir")
@@ -30,9 +42,7 @@ func TestCanonical_ExecOnlyDir(t *testing.T) {
 }
 
 func TestWeaklyCanonical_ExecOnlyDir(t *testing.T) {
-	if os.Getuid() == 0 {
-		t.Skip("root bypasses permission checks")
-	}
+	needs.Need(t, needs.NonRoot)
 
 	tmp := canonicalTempDir(t)
 	dir := filepath.Join(tmp, "execonly")
@@ -41,6 +51,7 @@ func TestWeaklyCanonical_ExecOnlyDir(t *testing.T) {
 
 	require.NoError(t, os.Chmod(dir, 0o111))
 	t.Cleanup(func() { _ = os.Chmod(dir, 0o755) })
+	requireExecOnlyEnforced(t, dir)
 
 	got, err := WeaklyCanonical(filepath.Join(dir, "file"))
 	require.NoError(t, err, "WeaklyCanonical must work with execute-only parent dir")
@@ -49,9 +60,7 @@ func TestWeaklyCanonical_ExecOnlyDir(t *testing.T) {
 
 func TestCanonical_ExecOnlyDirCaseNormalization(t *testing.T) {
 	skipIfCaseSensitive(t)
-	if os.Getuid() == 0 {
-		t.Skip("root bypasses permission checks")
-	}
+	needs.Need(t, needs.NonRoot)
 
 	tmp := canonicalTempDir(t)
 	dir := filepath.Join(tmp, "ExecOnly")
@@ -60,6 +69,7 @@ func TestCanonical_ExecOnlyDirCaseNormalization(t *testing.T) {
 
 	require.NoError(t, os.Chmod(dir, 0o111))
 	t.Cleanup(func() { _ = os.Chmod(dir, 0o755) })
+	requireExecOnlyEnforced(t, dir)
 
 	// On a case-insensitive FS, asking for wrong-case "mixedcase" should
 	// still return the on-disk name "MixedCase".
